This is now not really the RoS submission, it is a complete reformat of the item system :)
