+++ bots/behaviorLib.lua
@@ -2151,433 +2151,6 @@ behaviorLib.DontBreakChannelBehavior["Name"] = "DontBreakChannel"
 tinsert(behaviorLib.tBehaviors, behaviorLib.DontBreakChannelBehavior)
 
 
---------------------------------------
---          UseHealthRegen          --
---------------------------------------
---
--- Utility: 0 to 40
--- Based on missing health
---
--- Execute: 
--- Use a Rune of the Blight, Health Pot, or Bottle to heal
--- Will only use Health Pot or Bottle if it is safe
---
-
--------- Global Constants & Variables --------
-behaviorLib.bUseBatterySupplyForHealth = true
-behaviorLib.bUseBatterySupplyForMana = true
-behaviorLib.bUseBottleForHealth = true
-behaviorLib.bUseBottleForMana = true
-
-behaviorLib.safeTreeAngle = 120
-
--------- Helper Functions --------
-function behaviorLib.GetSafeDrinkDirection()
-	-- Returns vector to a safe direciton to retreat to drink if the bot is threatened
-	-- Returns nil if safe
-	local vecSafeDirection = nil
-	local unitSelf = core.unitSelf
-	local vecSelfPos = unitSelf:GetPosition()
-	local nMyID = unitSelf:GetUniqueID()
-	local tThreateningUnits = {}
-	local tUnitThreatenedRadius = {}
-	
-	for _, unitEnemy in pairs(core.localUnits["EnemyUnits"]) do
-		-- Ignore creeps that are already attacking something
-		local unitEnemyTarget = unitEnemy:GetAttackTarget()
-		if not (core.IsLaneCreep(unitEnemy) and unitEnemyTarget and unitEnemyTarget:GetUniqueID() ~= nMyID) then
-			local nAbsRange = core.GetAbsoluteAttackRangeToUnit(unitEnemy, unitSelf) + 325
-			local nAbsRangeSq = nAbsRange * nAbsRange
-			local nDistSq = Vector3.Distance2DSq(vecSelfPos, unitEnemy:GetPosition())
-			if nDistSq < nAbsRangeSq then
-				tinsert(tThreateningUnits, unitEnemy)
-				tinsert(tUnitThreatenedRadius, nAbsRange)
-			end
-		end
-	end
-
-	local curTimeMS = HoN.GetGameTime()
-	local nThreateningUnits = core.NumberElements(tThreateningUnits)
-	if nThreateningUnits > 0 or eventsLib.recentDotTime > curTimeMS or #eventsLib.incomingProjectiles["all"] > 0 then
-		-- Determine best "away from threat" vector
-		local vecAway = Vector3.Create()
-		for nIndex, unitEnemy in pairs(tThreateningUnits) do
-			local vecAwayFromTarget = Vector3.Normalize(vecSelfPos - unitEnemy:GetPosition())
-			vecAway = vecAway + vecAwayFromTarget * tUnitThreatenedRadius[nIndex]
-		end
-		
-		-- Average vecAway with "retreat" vector
-		local vecRetreat = Vector3.Normalize(behaviorLib.PositionSelfBackUp() - vecSelfPos)
-		local vecSafeDirection = Vector3.Normalize(vecAway + vecRetreat)
-	end
-
-	return vecSafeDirection
-end
-
-------------------------------------
---   Mana Battery/PowerSupply     --
-------------------------------------
-
-function behaviorLib.GetBatterySupplyFromInventory()
-	-- Returns Mana Battery or Power Supply if they are in the bot's inventory
-	-- else returns nil
-	
-	local unitSelf = core.unitSelf
-	local tInventory = unitSelf:GetInventory()
-	local tManaBattery = core.InventoryContains(tInventory, "Item_ManaBattery")
-	local tPowerSupply = core.InventoryContains(tInventory, "Item_PowerSupply")
-	
-	if not core.IsTableEmpty(tManaBattery) then
-		return tManaBattery[1]
-	elseif not core.IsTableEmpty(tPowerSupply) then
-		return tPowerSupply[1]
-	end
-	
-	return nil
-end
-
-function behaviorLib.BatterySupplyHealthUtilFn(nHealthMissing, nCharges)
-	-- With 1 Charge:
-	-- Roughly 20+ when we are missing 30 health
-	-- Function which crosses 20 at x=30 and 30 at x=140, convex down
-	-- With 15 Charges:
-	-- Roughly 20+ when we are missing 170 health
-	-- Function which crosses 20 at x=170 and 30 at x=330, convex down
-	if (not behaviorLib.bUseBatterySupplyForHealth) then
-		return 0
-	end
-	
-	local nHealAmount = 10 * nCharges
-	local nHealBuffer = 20
-	local nUtilityThreshold = 20
-	
-	local vecPoint = Vector3.Create(nHealAmount + nHealBuffer, nUtilityThreshold)
-	local vecOrigin = Vector3.Create(-250, -30)
-	
-	return core.ATanFn(nHealthMissing, vecPoint, vecOrigin, 100)
-end
-
-function behaviorLib.BatterySupplyManaUtilFn(nManaMissing, nCharges)
-	-- With 1 Charge:
-	-- Roughly 20+ when we are missing 40 mana
-	-- Function which crosses 20 at x=40 and 30 at x=100, convex down
-	-- With 15 Charges:
-	-- Roughly 20+ when we are missing 280 mana
-	-- Function which crosses 20 at x=280 and 30 at x=470, convex down
-	if (not behaviorLib.bUseBatterySupplyForMana) then
-		return 0
-	end
-	
-	local nManaRegenAmount = 15 * nCharges
-	local nManaBuffer = 25
-	local nUtilityThreshold = 20
-	
-	local vecPoint = Vector3.Create(nManaRegenAmount + nManaBuffer, nUtilityThreshold)
-	local vecOrigin = Vector3.Create(-60, -50)
-	
-	return core.ATanFn(nManaMissing, vecPoint, vecOrigin, 100)
-end
-
-function behaviorLib.UseBatterySupplyUtility(botBrain)
-	local unitSelf = core.unitSelf
-	local nHealthMissing = unitSelf:GetMaxHealth() - unitSelf:GetHealth()
-	local nManaMissing = unitSelf:GetMaxMana() - unitSelf:GetMana()
-	local itemBatterySupply = behaviorLib.GetBatterySupplyFromInventory()
-	
-	if itemBatterySupply ~= nil and itemBatterySupply:CanActivate() then
-		local nCharges = itemBatterySupply:GetCharges()
-		local nBatterySupplyHealthUtility = behaviorLib.BatterySupplyHealthUtilFn(nHealthMissing, nCharges)
-		local nBatterySupplyManaUtility = behaviorLib.BatterySupplyManaUtilFn(nManaMissing, nCharges)
-		
-		return max(
-			nBatterySupplyHealthUtility * .8 + nBatterySupplyManaUtility * .2, --health
-			nBatterySupplyManaUtility * .8 + nBatterySupplyHealthUtility * .2  --mana
-			)
-	end
-	
-	return 0
-end
-
-function behaviorLib.UseBatterySupplyExecute(botBrain)
-	local unitSelf = core.unitSelf
-	local tInventory = unitSelf:GetInventory()
-	local bActionTaken = false
-	
-	-- Use Mana Battery/Power Supply to heal
-	local itemBatterySupply = behaviorLib.GetBatterySupplyFromInventory(tInventory)
-	if itemBatterySupply:GetCharges() > 0 then
-		bActionTaken = core.OrderItemClamp(botBrain, unitSelf, itemBatterySupply)
-	end
-	
-	return bActionTaken
-end
-behaviorLib.UseBatterySupplyBehavior = {}
-behaviorLib.UseBatterySupplyBehavior["Utility"] = behaviorLib.UseBatterySupplyUtility
-behaviorLib.UseBatterySupplyBehavior["Execute"] = behaviorLib.UseBatterySupplyExecute
-behaviorLib.UseBatterySupplyBehavior["Name"] = "UseBatterySupply"
-tinsert(behaviorLib.tBehaviors, behaviorLib.UseBatterySupplyBehavior)
-
-------------------------------------
---       Runes OfThe Blight       --
-------------------------------------
-function behaviorLib.UseRunesOfTheBlightUtility(botBrain)
-	-- Roughly 20+ when we are missing 115 hp
-	-- Function which crosses 20 at x=115 and is 30 at roughly x=600, convex down
-
-	local unitSelf = core.unitSelf
-	local tInventory = unitSelf:GetInventory()
-	local tBlights = core.InventoryContains(tInventory, "Item_RunesOfTheBlight")
-	
-	if not core.IsTableEmpty(tBlights) and not unitSelf:HasState("State_RunesOfTheBlight") then
-		local unitSelf = core.unitSelf
-		local nHealthMissing = unitSelf:GetMaxHealth() - unitSelf:GetHealth()
-		local nHealthRegen = unitSelf:GetHealthRegen()
-		local nHealAmount = 115
-		local nHealBuffer = nHealthRegen * 16
-		local nUtilityThreshold = 20
-			
-		local vecPoint = Vector3.Create(nHealAmount + nHealBuffer, nUtilityThreshold)
-		local vecOrigin = Vector3.Create(-1000, -20)
-		
-		return core.ATanFn(nHealthMissing, vecPoint, vecOrigin, 100)
-	end
-	
-	return 0
-end
-
-function behaviorLib.UseRunesOfTheBlightExecute(botBrain)
-	local unitSelf = core.unitSelf
-	local vecSelfPos = unitSelf:GetPosition()
-	local tInventory = unitSelf:GetInventory()
-	local tBlights = core.InventoryContains(tInventory, "Item_RunesOfTheBlight")
-	local bActionTaken = false
-	local unitClosestTree = nil
-	local nClosestTreeDistSq = 9999 * 9999
-	local vecLaneForward = object.vecLaneForward
-	local vecLaneForwardNeg = -vecLaneForward
-	local funcRadToDeg = core.RadToDeg
-	local funcAngleBetween = core.AngleBetween
-	local nHalfSafeTreeAngle = behaviorLib.safeTreeAngle / 2
-
-	if not core.IsTableEmpty(tBlights) then
-		core.UpdateLocalTrees()
-		local tTrees = core.localTrees
-		for _, unitTree in pairs(tTrees) do
-			vecTreePosition = unitTree:GetPosition()
-			-- "Safe" trees are backwards
-			if not vecLaneForward or abs(funcRadToDeg(funcAngleBetween(vecTreePosition - vecSelfPos, vecLaneForwardNeg)) ) < nHalfSafeTreeAngle then
-				local nDistSq = Vector3.Distance2DSq(vecTreePosition, vecSelfPos)
-				if nDistSq < nClosestTreeDistSq then
-					unitClosestTree = unitTree
-					nClosestTreeDistSq = nDistSq
-					if bDebugLines then
-						core.DrawXPosition(vecTreePosition, 'yellow')
-					end
-				end
-			end
-		end
-		if unitClosestTree ~= nil then
-			bActionTaken = core.OrderItemEntityClamp(botBrain, unitSelf, tBlights[1], unitClosestTree)
-		end
-	end
-		
-	return bActionTaken
-end
-behaviorLib.UseRunesOfTheBlightBehavior = {}
-behaviorLib.UseRunesOfTheBlightBehavior["Utility"] = behaviorLib.UseRunesOfTheBlightUtility
-behaviorLib.UseRunesOfTheBlightBehavior["Execute"] = behaviorLib.UseRunesOfTheBlightExecute
-behaviorLib.UseRunesOfTheBlightBehavior["Name"] = "UseRunesOfTheBlight"
-tinsert(behaviorLib.tBehaviors, behaviorLib.UseRunesOfTheBlightBehavior)
-
-------------------------------------
---          Health Potion         --
-------------------------------------
-function behaviorLib.UseHealthPotUtility(botBrain)
-	-- Roughly 20+ when we are missing 400 hp
-	-- Function which crosses 20 at x=400 and 40 at x=650, convex down
-
-	local unitSelf = core.unitSelf
-	local tInventory = unitSelf:GetInventory()
-	local tHealthPots = core.InventoryContains(tInventory, "Item_HealthPotion")
-	
-	if not core.IsTableEmpty(tHealthPots) and not unitSelf:HasState("State_HealthPotion") then	
-		local nHealthMissing = unitSelf:GetMaxHealth() - unitSelf:GetHealth()
-		local nHealthRegen = unitSelf:GetHealthRegen()
-		local nHealAmount = 400
-		local nHealBuffer = nHealthRegen * 10
-		local nUtilityThreshold = 20
-		
-		local vecPoint = Vector3.Create(nHealAmount, nUtilityThreshold)
-		local vecOrigin = Vector3.Create(200, -40)
-		return core.ATanFn(nHealthMissing, vecPoint, vecOrigin, 100)
-	end
-	
-	return 0
-end
-
-function behaviorLib.UseHealthPotExecute(botBrain)
-	local bActionTaken = false
-	local unitSelf = core.unitSelf
-	local vecSelfPos = unitSelf:GetPosition()
-	local tInventory = unitSelf:GetInventory()
-	local tHealthPots = core.InventoryContains(tInventory, "Item_HealthPotion")
-	
-	if not core.IsTableEmpty(tHealthPots) then
-		local vecRetreatDirection = behaviorLib.GetSafeDrinkDirection()
-		-- Check if it is safe to drink
-		if vecRetreatDirection then
-			bActionTaken = core.OrderMoveToPosClamp(botBrain, unitSelf, vecSelfPos + vecRetreatDirection * core.moveVecMultiplier, false)
-		else
-			bActionTaken = core.OrderItemEntityClamp(botBrain, unitSelf, tHealthPots[1], unitSelf)
-		end
-	end
-	
-	return bActionTaken
-end
-
-behaviorLib.UseHealthPotBehavior = {}
-behaviorLib.UseHealthPotBehavior["Utility"] = behaviorLib.UseHealthPotUtility
-behaviorLib.UseHealthPotBehavior["Execute"] = behaviorLib.UseHealthPotExecute
-behaviorLib.UseHealthPotBehavior["Name"] = "UseHealthPot"
-tinsert(behaviorLib.tBehaviors, behaviorLib.UseHealthPotBehavior)
-
-------------------------------------
---             Bottle             --
-------------------------------------
-function behaviorLib.BottleHealthUtilFn(nHealthMissing, nHealthRegen)
-	-- Roughly 20+ when we are missing 135 hp
-	-- Function which crosses 20 at x=135 and 30 at x=220, convex down
-	if (not behaviorLib.bUseBottleForHealth) then
-		return 0
-	end
-	
-	local nHealAmount = 135
-	local nHealBuffer = nHealthRegen * 3
-	local nUtilityThreshold = 20
-
-	local vecPoint = Vector3.Create(nHealAmount + nHealBuffer, nUtilityThreshold)
-	local vecOrigin = Vector3.Create(-100, -30)
-	
-	return core.ATanFn(nHealthMissing, vecPoint, vecOrigin, 100)
-end
-
-function behaviorLib.BottleManaUtilFn(nManaMissing, nManaRegen)
-	-- Roughly 20+ when we are missing 70 mana
-	-- Function which crosses 20 at x=70 and 30 at x=140, convex down
-	if (not behaviorLib.bUseBottleForMana) then
-		return 0
-	end
-	
-	local nManaRegenAmount = 70
-	local nManaBuffer = nManaRegen * 3
-	local nUtilityThreshold = 20
-	
-	local vecPoint = Vector3.Create(nManaRegenAmount + nManaBuffer, nUtilityThreshold)
-	local vecOrigin = Vector3.Create(-125, -30)
-	
-	return core.ATanFn(nManaMissing, vecPoint, vecOrigin, 100)
-end
-
-function behaviorLib.UseBottleUtility(botBrain)
-	local unitSelf = core.unitSelf
-	local nHealthMissing = unitSelf:GetMaxHealth() - unitSelf:GetHealth()
-	local nManaMissing = unitSelf:GetMaxMana() - unitSelf:GetMana()
-	local nHealthRegen = unitSelf:GetHealthRegen()
-	local nManaRegen = unitSelf:GetManaRegen()
-	local tInventory = unitSelf:GetInventory()
-	local tItemBottle = core.InventoryContains(tInventory, "Item_Bottle")
-	
-	if not core.IsTableEmpty(tItemBottle) and not unitSelf:HasState("State_Bottle") and tItemBottle[1]:GetActiveModifierKey() ~= "bottle_empty" then
-		local nBottleHealthFn=behaviorLib.BottleHealthUtilFn(nHealthMissing, nHealthRegen)
-		local nBottleManaFn=behaviorLib.BottleManaUtilFn(nManaMissing, nManaRegen)
-		
-		return max(
-			nBottleHealthFn * .8 + nBottleManaFn * .2, --health
-			nBottleManaFn * .8 + nBottleHealthFn * .2  --mana
-			)
-	end
-	
-	return 0
-end
-
-function behaviorLib.UseBottleExecute(botBrain)
-	local bActionTaken = false
-	local unitSelf = core.unitSelf
-	local tInventory = unitSelf:GetInventory()
-	local tItemBottle = core.InventoryContains(tInventory, "Item_Bottle")
-
-	if not core.IsTableEmpty(tItemBottle) then
-		local vecRetreatDirection = behaviorLib.GetSafeDrinkDirection()
-		-- Check if it is safe to drink
-		if vecRetreatDirection then
-			bActionTaken = core.OrderMoveToPosClamp(botBrain, unitSelf, vecSelfPos + vecRetreatDirection * core.moveVecMultiplier, false)
-		else
-			bActionTaken = core.OrderItemClamp(botBrain, unitSelf, tItemBottle[1])
-		end
-	end
-	
-	return bActionTaken
-end
-
-behaviorLib.UseBottleBehavior = {}
-behaviorLib.UseBottleBehavior["Utility"] = behaviorLib.UseBottleUtility
-behaviorLib.UseBottleBehavior["Execute"] = behaviorLib.UseBottleExecute
-behaviorLib.UseBottleBehavior["Name"] = "UseBottle"
-tinsert(behaviorLib.tBehaviors, behaviorLib.UseBottleBehavior)
-
-------------------------------------
---          Mana Potion           --
-------------------------------------
-function behaviorLib.UseManaPotUtility(botBrain)
-	-- Roughly 20+ when we are missing 100 mana
-	-- Function which crosses 20 at x=100 and 30 at x=200, convex down
-	
-	local unitSelf = core.unitSelf
-	local tInventory = unitSelf:GetInventory()
-	local tManaPots = core.InventoryContains(tInventory, "Item_ManaPotion")
-	
-	if not core.IsTableEmpty(tManaPots) and not unitSelf:HasState("State_ManaPotion") then	
-		local nManaMissing = unitSelf:GetMaxMana() - unitSelf:GetMana()
-		local nManaRegen = unitSelf:GetManaRegen()
-		local nManaRegenAmount = 100
-		local nManaBuffer = nManaRegen * 20
-		local nUtilityThreshold = 20
-		
-		local vecPoint = Vector3.Create(nManaRegenAmount, nUtilityThreshold)
-		local vecOrigin = Vector3.Create(-100, -45)
-		
-		return core.ATanFn(nManaMissing, vecPoint, vecOrigin, 100)
-	end
-	
-	return 0
-end
-
-function behaviorLib.UseManaPotExecute(botBrain)
-	local bActionTaken = false
-	local unitSelf = core.unitSelf
-	local vecSelfPos = unitSelf:GetPosition()
-	local tInventory = unitSelf:GetInventory()
-	local tManaPots = core.InventoryContains(tInventory, "Item_ManaPotion")
-	
-	if not core.IsTableEmpty(tManaPots) then
-		local vecRetreatDirection = behaviorLib.GetSafeDrinkDirection()
-		-- Check if it is safe to drink
-		if vecRetreatDirection then
-			bActionTaken = core.OrderMoveToPosClamp(botBrain, unitSelf, vecSelfPos + vecRetreatDirection * core.moveVecMultiplier, false)
-		else
-			bActionTaken = core.OrderItemEntityClamp(botBrain, unitSelf, tManaPots[1], unitSelf)
-		end
-	end
-		
-	return bActionTaken
-end
-
-behaviorLib.UseManaPotBehavior = {}
-behaviorLib.UseManaPotBehavior["Utility"] = behaviorLib.UseManaPotUtility
-behaviorLib.UseManaPotBehavior["Execute"] = behaviorLib.UseManaPotExecute
-behaviorLib.UseManaPotBehavior["Name"] = "UseManaPot"
-tinsert(behaviorLib.tBehaviors, behaviorLib.UseManaPotBehavior)
-
 ----------------------------------
 --	PositionSelf
 --
@@ -3601,3 +3174,6 @@ behaviorLib.ShopBehavior["Utility"] = behaviorLib.ShopUtility
 behaviorLib.ShopBehavior["Execute"] = behaviorLib.ShopExecute
 behaviorLib.ShopBehavior["Name"] = "Shop"
 tinsert(behaviorLib.tBehaviors, behaviorLib.ShopBehavior)
+
+--add the item behaviors into the game, so they can be overloaded.
+runfile "bots/itemBehaviors.lua"
