+++ bots/core.lua
@@ -141,7 +141,7 @@ function core.CoreInitialize(controller)
 			core.printTableTable(tSorted)
 		end
 	end
-
+	
 	core.coreInitialized = true
 end
 
@@ -1172,6 +1172,24 @@ function core.InventoryContains(inventory, val, bIgnoreRecipes, bIncludeStash)
 	end
 	
     return tableOfThings
+end
+
+--Finds an item on your hero.
+function core.GetItem(val, bIncludeStash)
+	inventory = core.unitSelf:GetInventory()
+	if bIncludeStash == nil then
+		bIncludeStash = false
+	end
+	local nLast = (bIncludeStash and 12) or 6
+	for slot = 1, nLast, 1 do
+		local curItem = inventory[slot]
+		if curItem then
+			if curItem:GetName() == val then
+				return core.WrapInTable(curItem)
+			end
+		end
+	end
+    return nil
 end
 
 function core.IsLaneCreep(unit)
